feat(host): refresh config on provider restart
