+++ crates/host/src/wasmbus/providers/http_server/mod.rs
@@ -1,7 +1,6 @@
 use core::net::SocketAddr;
 use core::str::FromStr as _;
 
-use std::collections::HashMap;
 use std::future::Future;
 use std::sync::Arc;
 
@@ -11,7 +10,7 @@ use nkeys::XKey;
 use tokio::sync::{broadcast, Mutex};
 use tokio::task::JoinSet;
 use tracing::{error, instrument};
-use wasmcloud_core::InterfaceLinkDefinition;
+use wasmcloud_core::HostData;
 use wasmcloud_provider_http_server::default_listen_address;
 use wasmcloud_provider_sdk::provider::{
     handle_provider_commands, receive_link_for_provider, ProviderCommandReceivers,
@@ -31,21 +30,21 @@ impl crate::wasmbus::Host {
     #[instrument(level = "debug", skip_all)]
     pub(crate) async fn start_http_server_provider(
         &self,
-        link_definitions: impl IntoIterator<Item = InterfaceLinkDefinition>,
+        host_data: HostData,
         provider_xkey: XKey,
-        host_config: HashMap<String, String>,
         provider_id: &str,
         host_id: &str,
     ) -> anyhow::Result<JoinSet<()>> {
         let mut tasks = JoinSet::new();
-        let default_address = host_config
+        let default_address = host_data
+            .config
             .get("default_address")
             .map(|s| SocketAddr::from_str(s))
             .transpose()
             .context("failed to parse default_address")?
             .unwrap_or_else(default_listen_address);
 
-        let provider = match host_config.get("routing_mode").map(String::as_str) {
+        let provider = match host_data.config.get("routing_mode").map(String::as_str) {
             // Run provider in address mode by default
             Some("address") | None => HttpServerProvider::Address(address::Provider {
                 address: default_address,
@@ -82,7 +81,7 @@ impl crate::wasmbus::Host {
             Arc::from(provider_id),
             Arc::clone(&self.host_config.lattice),
             host_id.to_string(),
-            host_config,
+            host_data.config,
             provider_xkey,
             Arc::clone(&self.secrets_xkey),
         )
@@ -90,7 +89,7 @@ impl crate::wasmbus::Host {
 
         match provider {
             HttpServerProvider::Address(provider) => {
-                for ld in link_definitions {
+                for ld in host_data.link_definitions {
                     if let Err(e) = receive_link_for_provider(&provider, &conn, ld).await {
                         error!(
                             error = %e,
@@ -104,7 +103,7 @@ impl crate::wasmbus::Host {
                 });
             }
             HttpServerProvider::Path(provider) => {
-                for ld in link_definitions {
+                for ld in host_data.link_definitions {
                     if let Err(e) = receive_link_for_provider(&provider, &conn, ld).await {
                         error!(
                             error = %e,
